adidt: parts: hmc7044: fix pll2-output-frequency setting
We should not rely on the first position of the clocks configuration to
*really* have the PLL frequency we want to set. There might be
applications (eg: synchrona) where only a subset of the 14 channels are
enabled and pyadi-jif [1] is used to get the clock configuration to use.
Pyadi-jif will only return a list with the enabled channels, but,
nevertheless, apps have to provide the full 14 channel list to adidt which
means that if channel 0 is not to be enabled, it might not have trustworthy
info.

Since we already have the PLL frequency we want to set returned from
JIF, we should just use it. As a side effect, the code also looks much
simpler.

[1]: https://github.com/analogdevicesinc/pyadi-jif
Signed-off-by: Nuno Sá <nuno.sa@analog.com>